package: openssl: compiling error in bn-mips.S
fix the compilng error for mips32r6 in bn-mips.S
The error is as following:
crypto/bn/bn-mips.S:56: Error: opcode not supported on this processor: mips2 (mips2) `mulu $1,$12,$7'
so we make the patch to crypto/bn/asm/mips.pl. when isa is less than r6,
then the code ".set mips2" takes effect.

Signed-off-by: Hua Zhang <hua.zhang1974@hotmail.com>